+++ app/guests/routes.py
@@ -20,41 +20,45 @@ def index() -> str:
     return render_template("guests/_index.html")
 
 
-@blueprint.route("/best-of-only")
-def best_of_only() -> str:
-    """View: Guests Best Of Only Report."""
+@blueprint.route("/best-of-only-not-my-job-guests")
+def best_of_only_not_my_job_guests() -> str:
+    """View: Best Of Only Not My Job Guests Report."""
     _database_connection = mysql.connector.connect(**current_app.config["database"])
     _guests = retrieve_best_of_only_guests(database_connection=_database_connection)
     _database_connection.close()
-    return render_template("guests/best-of-only.html", guests=_guests)
+    return render_template("guests/best-of-only-not-my-job-guests.html", guests=_guests)
 
 
 @blueprint.route("/most-appearances")
 def most_appearances() -> str:
-    """View: Guests Most Appearances Report."""
+    """View: Most Appearances Report."""
     _database_connection = mysql.connector.connect(**current_app.config["database"])
     _guests = guest_multiple_appearances(database_connection=_database_connection)
     _database_connection.close()
     return render_template("guests/most-appearances.html", guests=_guests)
 
 
-@blueprint.route("/scoring-exceptions")
-def scoring_exceptions() -> str:
-    """View: Guests Scoring Exceptions Report."""
+@blueprint.route("/not-my-job-scoring-exceptions")
+def not_my_job_scoring_exceptions() -> str:
+    """View: Not My Job Scoring Exceptions Report."""
     _database_connection = mysql.connector.connect(**current_app.config["database"])
     _exceptions = retrieve_all_scoring_exceptions(
         database_connection=_database_connection
     )
     _database_connection.close()
-    return render_template("guests/scoring-exceptions.html", exceptions=_exceptions)
+    return render_template(
+        "guests/not-my-job-scoring-exceptions.html", exceptions=_exceptions
+    )
 
 
-@blueprint.route("/three-pointers")
-def three_pointers() -> str:
-    """View: Guests Three Pointers Report."""
+@blueprint.route("/not-my-job-three-pointers")
+def not_my_job_three_pointers() -> str:
+    """View: Not My Job Three Pointers Report."""
     _database_connection = mysql.connector.connect(**current_app.config["database"])
     _three_pointers = retrieve_all_three_pointers(
         database_connection=_database_connection
     )
     _database_connection.close()
-    return render_template("guests/three-pointers.html", three_pointers=_three_pointers)
+    return render_template(
+        "guests/not-my-job-three-pointers.html", three_pointers=_three_pointers
+    )