+++ app/main/routes.py
@@ -13,7 +13,7 @@ blueprint = Blueprint("main", __name__)
 
 @blueprint.route("/")
 def index() -> str:
-    """View: Site Index Page."""
+    """View: Site Index."""
     return render_template("pages/index.html")