+++ tests/test_guests.py
@@ -16,9 +16,9 @@ def test_index(client: FlaskClient) -> None:
     assert b"Best Of Only Not My Job Guests" in response.data
 
 
-def test_best_of_only(client: FlaskClient) -> None:
-    """Testing guests.routes.best_of_only."""
-    response: TestResponse = client.get("/guests/best-of-only")
+def test_best_of_only_not_my_job_guests(client: FlaskClient) -> None:
+    """Testing guests.routes.best_of_only_not_my_job_guests."""
+    response: TestResponse = client.get("/guests/best-of-only-not-my-job-guests")
     assert response.status_code == 200
     assert b"Best Of Only Not My Job Guests" in response.data
     assert b"Scoring Exception" in response.data
@@ -32,17 +32,17 @@ def test_most_appearances(client: FlaskClient) -> None:
     assert b"Regular" in response.data
 
 
-def test_scoring_exceptions(client: FlaskClient) -> None:
-    """Testing guests.routes.scoring_exceptions."""
-    response: TestResponse = client.get("/guests/scoring-exceptions")
+def test_not_my_job_scoring_exceptions(client: FlaskClient) -> None:
+    """Testing guests.routes.not_my_job_scoring_exceptions."""
+    response: TestResponse = client.get("/guests/not-my-job-scoring-exceptions")
     assert response.status_code == 200
     assert b"Not My Job Scoring Exceptions" in response.data
     assert b"Show Notes" in response.data
 
 
-def test_three_pointers(client: FlaskClient) -> None:
-    """Testing guests.routes.three_pointers."""
-    response: TestResponse = client.get("/guests/three-pointers")
+def test_not_my_job_three_pointers(client: FlaskClient) -> None:
+    """Testing guests.routes.not_my_job_three_pointers."""
+    response: TestResponse = client.get("/guests/not-my-job-three-pointers")
     assert response.status_code == 200
     assert b"Not My Job Three Pointers" in response.data
     assert b"Scoring Exception" in response.data